pyproject.toml
[tool.poetry]
name = "djangoninja"
version = "0.1.0"
description = ""
authors = ["KooHyunJung <ghj3160@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0"
django-ninja = "0.16.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
